refactor(delphi): extract umap_narrative stages into importable modules
Move the numbered stage scripts' logic into an importable stages/ package so
the orchestrator can call them in-process, while keeping each stage runnable
standalone from the CLI:

  501_calculate_comment_extremity.py -> stages/comment_extremity.py
  502_calculate_priorities.py        -> stages/comment_priorities.py
  700_datamapplot_for_layer.py       -> stages/datamapplot_layer.py

Each numbered script becomes a 3-6 line CLI shim (from stages.<name> import
main; sys.exit(main())). Every stage main() now accepts an optional argv so it
can be invoked in-process without mutating global sys.argv. run_pipeline.main
gains the same argv parameter; reset_conversation gains a cli(argv) wrapper
while keeping main(zid, rid) intact for existing callers/tests.

Content-preserving: no algorithm or output changes. test_501 updated to import
the moved module (patch targets follow the code). Full suite: 570 passed,
17 skipped, 47 xfailed (unchanged from baseline).
